requirements.txt
asgiref==3.6.0
bleach==6.0.0
Django==4.2
django-summernote==0.8.20.0
mysqlclient==2.1.1
Pillow==9.5.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
webencodings==0.5.1
